pyproject.toml
[project]
name = "auto-ru-parser"
version = "0.1.1"
description = "Simple parser for auto data extraction"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "flake8>=7.2.0",
    "isort>=6.0.1",
    "mypy>=1.15.0",
    "ruff>=0.11.4",
    "types-requests>=2.32.0.20250328",
]
